+++ ConexaoCaninaApp/ConexaoCaninaApp.API/Program.cs
@@ -1,5 +1,7 @@
 using ConexaoCaninaApp.Infra.Data.Context;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -14,6 +16,17 @@ builder.Services.AddSwaggerGen();
 builder.Services.AddDbContext<ApplicationDbContext>(options =>
         options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
 
+builder.Services.AddIdentity<IdentityUser, IdentityRole>()
+    .AddEntityFrameworkStores<ApplicationDbContext>()
+    .AddDefaultTokenProviders();
+
+
+builder.Services.AddAuthorization(options =>
+{
+    options.AddPolicy("RegisteredUserOnly", policy =>
+    policy.RequireAuthenticatedUser());
+});
+
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
@@ -26,8 +39,11 @@ if (app.Environment.IsDevelopment())
 // MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;
 app.UseHttpsRedirection();
 
+app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapControllers();
 
 app.Run();
+
+public partial class Program { }
